pr-blocksのリファクタリング
